Making a little progress with py3 compatability

+++ peepdf/JSAnalysis.py
@@ -199,7 +199,7 @@ def isJavascript(content):
     if re.findall(reJSscript, content, re.DOTALL | re.IGNORECASE):
         return True
 
-    _, count = re.subn(bad_chars_re, "", content, len(content) / 10)
+    _, count = re.subn(bad_chars_re, "", content, int(len(content) / 10))
     if count == len(content) / 10:
         return False
